refactor: split composer and collect action
